Changed STATUS file-type

+++ 0x15-file_io/1-create_file.c
@@ -11,7 +11,7 @@ int create_file(const char *filename, char *text_content)
 {
 	int status;
 	size_t len = 0;
-	ssize_t write_f;
+	int write_f;
 
 	if (filename == NULL)
 		return (-1);
